Create extractor API package structure to ensure testable code
The extraction package is separated in to three sub-packages: api (contains
the code that users will interact with to manage extractors), config, and
extractors (implementation of the Influx Data Extractor).

Users will interract with the API by obtaining an instance of the API interface,
not through functions directly. The implementation of the interface is in a
separate file.

The API implementation creates an unexported struct which implements the
interface. This struct also defines all the external functions that the
implementation calls, allowing them to be mocked from outside.