Create an implementation for the Influx Extractor
The Influx Extractor uses the schema discovery lib to retreive the
column definition of a Influx measurement and then procedes to pull
the data in chunks and pass it to a Go channel. If an error occurs the
data channel is closed and an error is produced on a separate channel.