Implement proper error handling and execution pipelines (#5)
The extractors and ingestors are designed to be run in goroutines. In order to notify one of them that an error has occurred error handling via a Broadcast mechanism is implemented. Each goroutine can subscribe to the broadcaster and choose how to handle external errors. Also each goroutine can broadcast it's errors to other subscribers.

Additionally implemented is SchemaStrategy. The ingestor can be called with a schema strategy, either to drop existing tables and recreate them, or just validate them or create only if missing