Rearrange the project structure
We are making a push to make our projects eventually all conform
with Golang best practices and this project setup seems to be widely
agreed upon in the community.

For now I have moved all packages into internal/; however, we may
find that some of them are useful to expose to other projects, at
which point we should move them to pkg/. All the files related to
the 'outflux' binary itself are now located in cmd/outflux.

A small bug fix is also included in schema_transfer.go, where the
wrong error message was being printed.