Remove pq and sql.DB, instead use pgx for all Timescale connections (#16)
When using the the pgx package we managed to increase the performance of outflux
when inserting data in timescale ~10k rows/second. This commit will replace the usage of the pq package together with the sql package so that only pgx is used throughout the code.
Additionally pgx simplifies the connection establishment and management.