Make pipelines used for schema transfer and migration (#26)
Previously schema-transfer was a separate action that needed to be explicitly done. To increase reusability of the code and allow for future expansion, the pipeline elements (extractor and ingestor) are redesigned to work in 2 steps, prepare and run. An Extractor will prepare by discovering the schema to be exported, and an Ingestor will prepare by creating/modifying the target db to accommodate the incoming data.